Bypass gomobile init latest gobind in CI

+++ buildScript/lib/core/get_source.sh
@@ -9,16 +9,27 @@ pushd ..
 
 ####
 
+DEFAULT_SING_BOX_REPO="https://github.com/MatsuriDayo/sing-box.git"
+SING_BOX_REPO="${SING_BOX_REPO:-$DEFAULT_SING_BOX_REPO}"
 if [ ! -d "sing-box" ]; then
-  git clone --no-checkout "${SING_BOX_REPO:-https://github.com/MatsuriDayo/sing-box.git}" sing-box
+  git clone --no-checkout "$SING_BOX_REPO" sing-box
 fi
 pushd sing-box
 git checkout "$COMMIT_SING_BOX"
-git restore --source "$COMMIT_SING_BOX" -- option/shadowsocks.go outbound/shadowsocks.go
-rm -f outbound/shadowsocks_tls_direct.go
-TLS_DIRECT_PATCH="$ROOT_DIR/buildScript/lib/core/patches/ss-tls-direct.patch"
-git apply --check "$TLS_DIRECT_PATCH" || exit 1
-git apply "$TLS_DIRECT_PATCH" || exit 1
+if [ -z "${SING_BOX_APPLY_PATCH:-}" ]; then
+  if [ "$SING_BOX_REPO" = "$DEFAULT_SING_BOX_REPO" ]; then
+    SING_BOX_APPLY_PATCH=1
+  else
+    SING_BOX_APPLY_PATCH=0
+  fi
+fi
+if [ "$SING_BOX_APPLY_PATCH" = "1" ]; then
+  git restore --source "$COMMIT_SING_BOX" -- option/shadowsocks.go outbound/shadowsocks.go
+  rm -f outbound/shadowsocks_tls_direct.go
+  TLS_DIRECT_PATCH="$ROOT_DIR/buildScript/lib/core/patches/ss-tls-direct.patch"
+  git apply --check "$TLS_DIRECT_PATCH" || exit 1
+  git apply "$TLS_DIRECT_PATCH" || exit 1
+fi
 popd
 
 ####
